Grandstream_HT814/requirements.txt
hexdump==3.3
pyserial==3.4
